+++ lib/add_widget.dart
@@ -1,4 +1,4 @@
-import 'package:caregiver_app/subwidgets/create_event_widget.dart';
+import 'package:caregiver_app/subwidgets/add/create_event_widget.dart';
 import 'package:caregiver_app/theme.dart';
 import 'package:flutter/material.dart';
