+++ lib/subwidgets/events/detail_dot_widget.dart
